Better checking if binaries exist

+++ scripts/install.sh
@@ -2,16 +2,14 @@
 
 set -ex
 
-if [ ! -d "$HOME/bin" ]; then
-    mkdir -p $HOME/bin
+mkdir -p $HOME/bin
 
-    if [ $TRAVIS_OS_NAME = linux ] && [ $TRAVIS_RUST_VERSION = stable ] ; then
-        pushd $HOME
-            wget https://github.com/SimonKagstrom/kcov/archive/v35.tar.gz
-            tar xzf v35.tar.gz
-        popd
+if [ $TRAVIS_OS_NAME = linux ] && [ $TRAVIS_RUST_VERSION = stable ] && [ ! -f $HOME/bin/kcov ] ; then
+    pushd $HOME
+        wget https://github.com/SimonKagstrom/kcov/archive/v35.tar.gz
+        tar xzf v35.tar.gz
 
-        pushd $HOME/kcov-35
+        pushd $kcov-35
             mkdir -p build
             pushd build
                 cmake -DCMAKE_INSTALL_PREFIX=$HOME/bin ..
@@ -19,8 +17,10 @@ if [ ! -d "$HOME/bin" ]; then
                 sudo make install
             popd
         popd
-    fi
+    popd
+fi
 
+if [ ! -f $HOME/bin/capnp ] ; then
     curl -O https://capnproto.org/capnproto-c++-0.6.1.tar.gz
     tar zxf capnproto-c++-0.6.1.tar.gz
     cd capnproto-c++-0.6.1
